Updating setup information
Ran through this on a VM (Linux) where I didn't have miniconda installed. Have tidied up a few bits (and editor has
wrapped lines to 120 characters so it looks more than it is) but found there was some tweaks to be made and
clarifications and I've removed the section on creating a dedicated directory for this work since that is at the start
of Episode 02 of the material.