Tweaks to lessons 2 and 4
After running the pilot @bobturneruk and I decided that there were too many ways of doing things being introduced and
decided to focus on the workflow of...

1. Create environment
2. Activate
3. Install packages
4. Deactivate

The first challenge in the second episode hadn't been updated to reflect this and was installing packages whilst
creating the environment (perfectly legitimate but we wanted to avoid introducing too many ways of doing things).

This addresses that, tweaking the challenge and moving the callout to the installation of packages.